# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GrantAI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: Get matched grants for a user
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of matches with grants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchWithGrant"
  /matches/generate:
    post:
      operationId: generateMatches
      tags: [matches]
      summary: Generate AI-powered grant matches for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateMatchesBody"
      responses:
        "200":
          description: Generated matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchWithGrant"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    User:
      type: object
      properties:
        id:
          type: integer
        orgName:
          type: string
        orgType:
          type: string
        mission:
          type: string
        location:
          type: string
        focusAreas:
          type: string
        fundingNeed:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orgName
        - orgType
        - mission
        - location
        - focusAreas
        - createdAt
    CreateUserBody:
      type: object
      properties:
        orgName:
          type: string
        orgType:
          type: string
        mission:
          type: string
        location:
          type: string
        focusAreas:
          type: string
        fundingNeed:
          type: ["integer", "null"]
      required:
        - orgName
        - orgType
        - mission
        - location
        - focusAreas
    UpdateUserBody:
      type: object
      properties:
        orgName:
          type: string
        orgType:
          type: string
        mission:
          type: string
        location:
          type: string
        focusAreas:
          type: string
        fundingNeed:
          type: ["integer", "null"]
    Grant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        funder:
          type: string
        amount:
          type: ["integer", "null"]
        deadline:
          type: ["string", "null"]
        eligibility:
          type: string
        focusArea:
          type: string
        location:
          type: string
        description:
          type: string
        website:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - funder
        - eligibility
        - focusArea
        - location
        - description
        - createdAt
...
    MatchWithGrant:
      type: object
      properties:
        userId:
          type: integer
        grantId:
          type: integer
        fitScore:
          type: integer
        winProbability:
          type: ["integer", "null"]
        reasoning:
          type: string
        createdAt:
          type: string
          format: date-time
        grant:
          $ref: "#/components/schemas/Grant"
      required:
        - userId
        - grantId
        - fitScore
        - reasoning
        - createdAt
        - grant
    GenerateMatchesBody:
      type: object
      properties:
        userId:
          type: integer
      required:
        - userId
...