# lib/db/src/schema/users.ts
import { pgTable, text, serial, integer, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const usersTable = pgTable("users", {
  id: serial("id").primaryKey(),
  orgName: text("org_name").notNull(),
  orgType: text("org_type").notNull(),
  mission: text("mission").notNull(),
  location: text("location").notNull(),
  focusAreas: text("focus_areas").notNull(),
  fundingNeed: integer("funding_need"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertUserSchema = createInsertSchema(usersTable).omit({ id: true, createdAt: true });
export type InsertUser = z.infer<typeof insertUserSchema>;
export type User = typeof usersTable.$inferSelect;